Rename to ResourceOperationMessage
